requirements.txt
Flask
python-dotenv
google-generativeai
waitress